server/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-limiter==3.5.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests-cache==1.2.0
openai==1.51.2
beautifulsoup4==4.12.3
lxml==5.1.0
redis==5.0.1
tenacity==8.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.9